Fix download of specific monasca-common version

+++ run_maven.sh
@@ -58,6 +58,14 @@ if [ $RUN_BUILD = "true" ]; then
     fi
 fi
 
+# Result artifact
+if [ $RC -eq 0 -a -r target/*-shaded.jar ]; then
+    ARTIFACT_JAR=$(ls -1rt target/*-shaded.jar | tail -1)
+    ARTIFACT_VER=$(awk -F'[<>]' '/version/ { print $3; exit }' pom.xml)
+    cp $ARTIFACT_JAR target/monasca-persister.jar
+    printf "\nmonasca-persister $ARTIFACT_VER generated at target/monasca-persister.jar\n\n"
+fi
+
 # Cleanup and exit
 rm -fr ${MVN_PKG}*
 exit $RC
